requirements.txt
arxiv==2.2.0
bertopic==0.16.4
fitz==0.0.1.dev2
json_repair==0.47.8
jsonlines==4.0.0
networkx==3.5
numpy==2.3.2
openai==1.99.6
pydantic==2.11.7
pymilvus==2.5.11
python-dotenv==1.1.1
redis==6.4.0
Requests==2.32.4
scikit_learn==1.7.1
sentence_transformers==4.1.0
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.1
Unidecode==1.4.0
